refactor + lombok
