+++ src/main/java/ru/itmo/javazolotaya/lesson2/service/TemperatureConverter.java
@@ -1,4 +1,4 @@
-package ru.itmo.javazolotaya.lesson2;
+package ru.itmo.javazolotaya.lesson2.service;
 
 public interface TemperatureConverter {
     double convert(double value);
