+++ src/main/java/ru/itmo/javazolotaya/lesson2/service/TemperatureService.java
@@ -1,4 +1,4 @@
-package ru.itmo.javazolotaya.lesson2;
+package ru.itmo.javazolotaya.lesson2.service;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
