+++ src/main/java/ru/itmo/javazolotaya/lesson2/service/impl/CelsiusToKelvinConverter.java
@@ -1,7 +1,7 @@
-package ru.itmo.javazolotaya.lesson2;
-
+package ru.itmo.javazolotaya.lesson2.service.impl;
 
 import org.springframework.stereotype.Component;
+import ru.itmo.javazolotaya.lesson2.service.TemperatureConverter;
 
 @Component("celsiusToKelvin")
 public class CelsiusToKelvinConverter implements TemperatureConverter {
