+++ src/main/java/ru/itmo/javazolotaya/lesson2/service/impl/KelvinToCelsiusConverter.java
@@ -1,6 +1,7 @@
-package ru.itmo.javazolotaya.lesson2;
+package ru.itmo.javazolotaya.lesson2.service.impl;
 
 import org.springframework.stereotype.Component;
+import ru.itmo.javazolotaya.lesson2.service.TemperatureConverter;
 
 @Component("kelvinToCelsius")
 public class KelvinToCelsiusConverter implements TemperatureConverter {
